Updated to reflect Pipeline changes + Failed Samples (#22)
* Updated to reflect Pipeline changes + Failed Samples

Pipeline now requires a sample-sheet as a parameter during object creation. Some changes were made to this klp.py to reflect this.
Job objects are now able to audit their results based on a provided list of sample ids and return a list of the missing ones.
klp.py will now write out the list of samples that failed to process, and at what stage did they fail processing.

* Updated test_klp to set proper age of testing run-dir.

The updated Pipeline class in sequence_processing_pipeline now properly checks for the age of supplied run-directories, and checks to see if they are in the proper age range defined by configuration.json. This fix will allow the unittest to pass.

* Set time after directory has been modified and before testing

* Removed obsolete parameter

* Failed-samples reporting changed

Failed-samples from a Job are now written to file immediately after audit() completes in order to monitor progress and in case the Pipeline fails before the information can be written out to disk.

* flake8

* Updated Qiita studies identified in 'touched_studies.tsv'

If qp-klp moved one or more Pipeline-generated output files into a Qiita study during operation, that study will be recorded in a new file 'touched_studies.tsv'.
Each study will also be presented with its associated project name and URLs to the detailed study description in both Qiita-RC and Qiita.

* sample-sheet filenames are now sanitized

Spaces in sample-sheet filenames are now replaced with underscores '_' before writing the file to disk. This prevents unpredictable behavior downstream.

* Using metapool to handle lane numbers.

Using metapool package to handle adding lane numbers to a sample-sheet and saving it to disk.

* clean-up

* bugfix

* bugfix

* bugfix

* bugfix

* bugfix

* cleanup setup.py

* Added test for sample-sheets with spaces in their names.

* failed-samples list now organized by project

* building test for failed_samples.json

* Building test for failed samples recorder

* iterate on test

* force print output

* test_failed_samples_recorder() impl

* flake8

* bugfix

* touched_studies test

* bugfix

* touched studies test

* bugfix

* Move new files into final_folder

* bugfix