Silently handle leading-zeros in tube_ids
This code will assume that if tube_ids are present in a study that they are also the source of values in the sample_name column of a sample-sheet. If it encounters a discrepancy between sample-names in the sample-sheet and those in the study, it will silently remove any leading zeroes from the sample-names in the sheet and rechecking before reporting an error.